discord_ai_tutor_free/requirements.txt
discord.py==2.3.2
google-generativeai==0.3.0
python-dotenv==1.0.0
psutil==5.9.8
httpx==0.27.0
Flask==2.3.3
